requirements.txt
Pillow
matplotlib
numpy
pybrain
deap
cherrypy